chore: move data files to 'data' folder
