+++ src/getBpe.ts
@@ -1,8 +1,11 @@
 import fs from 'fs'
 import path from 'path'
-import { dictZip, range } from '../utils'
+import { dictZip, range } from './utils'
 
-const bpeFile = fs.readFileSync(path.join(__dirname, './vocab.bpe'), 'utf8')
+const bpeFile = fs.readFileSync(
+  path.join(__dirname, '../data/vocab.bpe'),
+  'utf8',
+)
 const lines = bpeFile.split('\n')
 
 const bpeMerges = lines.slice(1, -1).map((x) =>
@@ -15,6 +18,6 @@ const bpeMerges = lines.slice(1, -1).map((x) =>
 const bpeRanks = dictZip(bpeMerges, range(0, bpeMerges.length))
 
 fs.writeFileSync(
-  path.join(__dirname, './bpeRanks.json'),
+  path.join(__dirname, '../data/bpeRanks.json'),
   JSON.stringify(bpeRanks),
 )
